Split core memory module
