+++ packages/core/src/memory/options.ts
@@ -1,0 +1,7 @@
+import type { MemoryOptions, ResolvedMemoryOptions } from "./types";
+
+export function resolveMemoryOptions(options: MemoryOptions = {}): ResolvedMemoryOptions {
+  return {
+    savePolicy: options.savePolicy ?? "message",
+  };
+}
